Added cat and Cassowary which inherit Animal. Put each class in its own file. Got rid of Dog class.
