+++ CS352-PA1/Cassowary .cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CS352_PA1
+{
+    public class Cassowary : Animal
+    {
+        public Cassowary(string name, int age, float weight) : base(name, age, "bakawk", weight)
+        {
+
+        }
+    }
+}
